Chapter02 now has working increment of vector of data. Class data done by Data pointer, need to change to derived class with concrete data in implementation file

+++ Cpp14/TestChapter02/TestChapter02.cpp
@@ -31,7 +31,7 @@ namespace TestChapter02
 
 			Assert::AreEqual(before.size(), after.size());
 
-			for (int idx = 0; idx < before.size(); ++idx) {
+			for (size_t idx = 0; idx < before.size(); ++idx) {
 				int low = before[idx];
 				int hi = after[idx];
 				Assert::AreEqual(low + 1, hi);
